Remove homeassistant.remote
