+++ homeassistant/components/notify/aws_lambda.py
@@ -15,7 +15,7 @@ from homeassistant.const import (
 from homeassistant.components.notify import (
     ATTR_TARGET, PLATFORM_SCHEMA, BaseNotificationService)
 import homeassistant.helpers.config_validation as cv
-from homeassistant.remote import JSONEncoder
+from homeassistant.helpers.json import JSONEncoder
 
 REQUIREMENTS = ['boto3==1.4.7']
