+++ homeassistant/components/notify/html5.py
@@ -132,17 +132,6 @@ def _load_config(filename):
     return {}
 
 
-class JSONBytesDecoder(json.JSONEncoder):
-    """JSONEncoder to decode bytes objects to unicode."""
-
-    # pylint: disable=method-hidden, arguments-differ
-    def default(self, obj):
-        """Decode object if it's a bytes object, else defer to base class."""
-        if isinstance(obj, bytes):
-            return obj.decode()
-        return json.JSONEncoder.default(self, obj)
-
-
 class HTML5PushRegistrationView(HomeAssistantView):
     """Accepts push registrations from a browser."""
