+++ tests/components/test_demo.py
@@ -7,7 +7,7 @@ import pytest
 
 from homeassistant.setup import async_setup_component
 from homeassistant.components import demo, device_tracker
-from homeassistant.remote import JSONEncoder
+from homeassistant.helpers.json import JSONEncoder
 
 
 @pytest.fixture(autouse=True)
